+++ src/Modules/Core/DBConstants/Content/ContentId.php
@@ -16,4 +16,5 @@ class ContentId
 	public const SUPPORT_FOODSHARING_PAGE_42 = 42;
 	public const PRIVACY_POLICY_CONTENT = 28;
 	public const PRIVACY_NOTICE_CONTENT = 64;
+	public const BROADCAST_MESSAGE = 51;
 }